Organizing Rust code a bit
